+++ assets/js/admin/jsModule.js
@@ -1,0 +1,9 @@
+/* eslint-disable-next-line no-undef */
+const $ = jQuery;
+
+const init = () => {
+  console.log('admin js module initialized...');
+  console.log('jquery', $);
+};
+
+export default init;